requirements.txt
Flask==2.3.3
Pillow==10.0.1
google-generativeai==0.3.2
Werkzeug==2.3.7
